biometric-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
face-recognition==1.3.0
pymongo==4.5.0
numpy>=1.21.0  # More flexible version
python-multipart==0.0.6
pydantic>=2.0.0
cryptography==41.0.7
pillow>=10.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
